chore: refactor slicers
