+++ lib/migration-chunks/index.js
@@ -2,9 +2,9 @@
 
 const _ = require('lodash');
 const sliceByContentTypeId = require('./slicers/content-type-id');
-const sliceByFieldDelete = require('./slicers/field-delete');
+const sliceByFieldDelete = require('./slicers/into-separate-chunk')('field/delete');
 const slicedByContentTransform = require('./slicers/content-transform');
-const sliceByFieldIdChange = require('./slicers/field-id-change');
+const sliceByFieldIdChange = require('./slicers/into-separate-chunk')('field/rename');
 
 module.exports = function migrationPlan (migrationSteps) {
   const uninterruptedSteps = sliceByContentTypeId(migrationSteps);
